Add option to link cuda statically (#4422)
Summary:
When linking to cuda dynamically, and the target does not have cuda installed, the user can not even start code with FAISS compiled with cuda. Therefore it could be useful to link the CUDA runtime/blas statically instead to avoid requiring CUDA on the users end.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4422

Reviewed By: mnorris11

Differential Revision: D77827762

Pulled By: ramilbakhshyiev

fbshipit-source-id: 7d4ca0320ad38d3b757882ea4fd69bb3c5c18474